fix: Split out server code from software-updates/index
